requirements.txt
# Streamlit App
streamlit>=1.38.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
PyPDF2>=3.0.0

# ML & Local RAG
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
transformers>=4.41.0
accelerate>=0.30.0

# LangChain Ecosystem (Pinned for stability)
# Using the stable 0.2.x series to avoid dependency resolution issues
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-huggingface>=0.0.3
langchain-mistralai>=0.1.0
langgraph>=0.1.0
